lazy load markdown and nabla
